Mailing List
Mail_initiation: Setting the plumbing for the Mail object of the class
phpmailer.

Mailing_list: Begins the prcoess of subscription to mailing list by
adding the data to database with status set to 0, and dispatches a
confirmation mail.

subscribe: Confirms the subscription, ater user clikcs the link sent,
and sets the status to 1 in the database, to indicate a confirmed
subscription.

//confirmation link simply contains a flag for subscribe/unsubscribe and
regno for the register Number of that individual (to identify that
person's records and modify them in db).